refactor : remove commented code
this code was removed because it is useless our code base. It is
considered as death code. Our system works without it.
Note that some of the removed code is just print statements used during
debugging to investigate the state of each entity.
Code related to battery level monitoring also removed. In fact, this
feature is not introduced. We will work on it on next pull requests.